Add files via upload
This is an activity log which has 3 clients, here you can append your exercise and diet regime. This activity log will record and track the data, and it can be retrieved whenever needed. The code is written in python. The author of this program is Rishib Chakraborty. The exercise and diet activity log can be used to track the number of calories consumed in each meal, and also track the amount of calories burnt after a workout session, making this a fitness tracking activity log that can be used by users who are trying to achieve a fitness goal, or by a trainer who wants to keep a track of his clients.