Add files via upload
This is inspired from the game that was released in 2013. The objective of the game is to guide a bird through an endless series of pipes without touching them. The player must try to survive by avoiding the obstacles. The game ends when the bird crashes. It is a simple yet challenging retro gameplay. Hope you enjoy