Add files via upload
The name of this game is Snake water gun. Author is Rishib Chakraborty and it is programmed on python. The rules are simple. The snake drinks the water, the gun shoots the snake, and the gun has no effect on the water. You get to play against the computer. There are limited amount of chances. The tally is counted. Try out your luck in this adductive game.