Add files via upload
This is a website created for a fitness studio using HTML and CSS. This webpage has functional buttons and is designed as per the desired modern requirements of a space that is comfortable for working out. We all know the importance of a modern day fitness studio that comes with good instructors and equipments. Author of this project is Rishib Chakraborty.