Add files via upload
This is the timeless rock paper scissors game that is developed with the help of python. The user gets to enter the number of rounds they want to play. The rules are simple: paper beats rock, rock beats scissor, scissor beats rock. Your score will be recorded as you challenge the computer. The author of this program is Rishib Chakraborty.